consolidate tests and use quick-fail sapi client
